Front end

+++ back_end_server/file_handler.py
@@ -25,14 +25,21 @@ class FileSystemLock:
     def has_lock(self, path):
         return path in self.file_locks
 
-    def create_and_acquire_write_lock(self, path):
+    def create_lock(self, path):  # PRIVATE. Do not use on its own (Doesnt release mutex)
         self.mutex.acquire()
         if self.has_lock(path):
             self.mutex.release()
             raise RuntimeError  # TODO create specific error
-        lock = ReadWriteLock()
-        lock.acquire_write()
-        self.file_locks[path] = lock
+        self.file_locks[path] = ReadWriteLock()
+
+    def create_and_acquire_read_lock(self, path):
+        self.create_lock(path)
+        self.file_locks[path].acquire_read()
+        self.mutex.release()
+
+    def create_and_acquire_write_lock(self, path):
+        self.create_lock(path)
+        self.file_locks[path].acquire_write()
         self.mutex.release()
 
     def delete_lock(self):
@@ -52,7 +59,12 @@ class FileHandler:
     @staticmethod
     def fetch_file(path):
         path = '.' + path
-        FileHandler.locks.aquire_read(path)
+        if not os.path.isfile(path):  # Check if file exists
+            raise IOError('File does not exist')
+        if FileHandler.locks.has_lock(path):
+            FileHandler.locks.aquire_read(path)
+        else:
+            FileHandler.locks.create_and_acquire_read_lock(path)
         file = open(path, 'r')
         content = file.read()
         file.close()
@@ -62,7 +74,9 @@ class FileHandler:
     @staticmethod
     def create_file(path, content):
         path = '.' + path
-        FileHandler.locks.create_and_acquire_write_lock(path)
+        if os.path.isfile(path):  # Check if file exists
+            raise RuntimeError('File already exists')
+        FileHandler.locks.create_and_acquire_write_lock(path)  # If file does not exist, create a lock for it and the file itself
         FileHandler.ensure_dir(path)
         file = open(path, 'w+')
         file.write(content)
